Refactoring chart files
